+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.20.4] - <SET_DATE>
+## [0.21.0] - 2020-02-29
 
 ### Added
 
+- Added support for training Conversational AI models.
 - Added `cuda_device` parameter to MultiLabelClassificationModel.
+
+### Fixed
+
+- Fixed bug in MultiModalClassificationModel when `num_labels` is not given.
 
 ## [0.20.3] - 2020-02-22
 
@@ -425,7 +430,9 @@ Model checkpoint is now saved for all epochs again.
 
 - This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
 
-[0.20.3]: https://github.com/ThilinaRajapakse/simpletransformers/compare/daf5ccd...HEAD
+[0.21.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/f484717...HEAD
+
+[0.20.3]: https://github.com/ThilinaRajapakse/simpletransformers/compare/daf5ccd...f484717
 
 [0.20.2]: https://github.com/ThilinaRajapakse/simpletransformers/compare/538b8fa...daf5ccd